backend/requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
firebase-admin==6.2.0
flask==3.1.1
flask-cors==6.0.0
python-dotenv==1.1.0
tensorflow==2.12.0
scikit-learn==1.2.2
pandas==2.0.3 